backend/requirements.txt
fastapi==0.128.4
uvicorn==0.40.0
python-multipart==0.0.20
numpy==2.2.2
Pillow==11.1.0
rasterio==1.4.3
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
